in the middle of converting the skills to a skillresolver

+++ Assets/_Custom/Creatures/Skills/SkillSO.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Skill", menuName = "Scriptable Object/Skill/New Skill")]
 public class SkillSO : ItemSO
 {
-    public float staminaCost;
+    public int staminaCost;
     public int targetDamage;
-    public float selfDamage;
-    public float selfHeal;
+    public int selfDamage;
+    public int selfHeal;
     public float attackRange;
 }
